refactor(node): extract proof_history config into its own file
